# docs/Worth_It_OpenAPI_v1.0.yaml
openapi: 3.1.0
info:
  title: Worth It API
  version: 1.0.0
  description: Internal REST API contract for the Worth It mobile app MVP. All monetary amounts use integer minor
    units plus an ISO 4217 currency code.
servers:
- url: https://api.worthit.app/v1
  description: Production
- url: https://staging-api.worthit.app/v1
  description: Staging
security:
- bearerAuth: []
...
paths:
...
  /inventory:
    get:
      summary: List premium inventory
      tags:
      - Inventory
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/InventoryItem'
                  next_cursor:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    post:
      summary: Add scan or manual item to inventory
      tags:
      - Inventory
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '403':
          description: Premium required
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                scan_id:
                  type: string
                  format: uuid
                title:
                  type: string
                cost_basis:
                  $ref: '#/components/schemas/Money'
                purchase_date:
                  type: string
                  format: date
      parameters:
      - $ref: '#/components/parameters/IdempotencyKey'
  /inventory/{inventory_id}:
    patch:
      summary: Update inventory item
      tags:
      - Inventory
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                listing_price:
                  $ref: '#/components/schemas/Money'
                listing_platform:
                  type: string
                notes:
                  type: string
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
    delete:
      summary: Delete inventory item
      tags:
      - Inventory
      responses:
        '204':
          description: Deleted
        '403':
          description: Premium required
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
  /inventory/{inventory_id}/sale:
    post:
      summary: Record a completed sale
      tags:
      - Inventory
      responses:
        '200':
          description: Success
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InventoryItem'
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limited
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - sale_price
              - sale_date
              properties:
                sale_price:
                  $ref: '#/components/schemas/Money'
                sale_date:
                  type: string
                  format: date
                marketplace_fees:
                  $ref: '#/components/schemas/Money'
                outbound_shipping:
                  $ref: '#/components/schemas/Money'
                other_costs:
                  $ref: '#/components/schemas/Money'
      parameters:
      - name: inventory_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - $ref: '#/components/parameters/IdempotencyKey'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    webhookSignature:
      type: apiKey
      in: header
      name: X-Webhook-Signature
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        minLength: 16
        maxLength: 128
...
  schemas:
    Money:
      type: object
      required:
      - amount_minor
      - currency
      properties:
        amount_minor:
          type: integer
          description: Amount in the currency minor unit, e.g. 1299 = $12.99.
        currency:
          type: string
          pattern: ^[A-Z]{3}$
    Error:
      type: object
      required:
      - error
      - request_id
      properties:
        error:
          type: object
          required:
          - code
          - message
          - retryable
          properties:
            code:
              type: string
            message:
              type: string
            retryable:
              type: boolean
            details:
              type: object
              additionalProperties: true
        request_id:
          type: string
        timestamp:
          type: string
          format: date-time
...
    InventoryItem:
      type: object
      required:
      - id
      - status
      - cost_basis
      - created_at
      properties:
        id:
          type: string
          format: uuid
        scan_id:
          type: string
          format: uuid
        status:
          type: string
          enum:
          - acquired
          - listed
          - sold
          - archived
        title:
          type: string
        cost_basis:
          $ref: '#/components/schemas/Money'
        listing_price:
          $ref: '#/components/schemas/Money'
        sale_price:
          $ref: '#/components/schemas/Money'
        realized_profit:
          $ref: '#/components/schemas/Money'
        created_at:
          type: string
          format: date-time